requirements.txt
absl-py==1.0.0
aiofiles==23.2.1
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.2
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
appnope==0.1.2
argcomplete==3.1.6
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
array-record==0.2.0
arrow==1.2.3
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.2.0
autograd==1.6.2
awscli==1.22.25
Babel==2.11.0
backcall==0.2.0
backoff==2.2.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
black==23.1.0
bleach==5.0.1
blessed==1.20.0
blinker==1.6.2
blosc2==2.0.0
boto3==1.29.4
botocore==1.32.4
Brotli==1.1.0
cachetools==4.2.4
cffi==1.15.1
charset-normalizer==2.0.9
click==8.1.3
clip==0.2.0
colorama==0.4.3
comm==0.1.2
contextlib2==21.6.0
contourpy==1.0.6
croniter==1.4.1
cycler==0.11.0
Cython==3.0.5
dacite==1.7.0
dateutils==0.6.12
debugpy==1.6.4
decorator==5.1.0
deepdiff==6.3.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
dm-tree==0.1.8
dnspython==2.4.2
docker-pycreds==0.4.0
docutils==0.15.2
einops==0.7.0
entrypoints==0.4
etils==1.2.0
exceptiongroup==1.2.0
fastapi==0.100.0
fastjsonschema==2.16.2
ffmpeg==1.4
ffmpeg-python==0.2.0
fiftyone==0.22.3
fiftyone-brain==0.13.3
fiftyone-db==0.4.0
filelock==3.12.0
fire==0.5.0
flatbuffers==23.3.3
fonttools==4.38.0
fqdn==1.5.1
frozendict==2.3.7
frozenlist==1.3.3
fsspec==2023.6.0
ftfy==6.1.1
functorch==2.0.0
future==0.18.2
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.32
glob2==0.7
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.59.0
graphql-core==3.2.3
graphviz==0.17
grpcio==1.54.0
h11==0.14.0
h2==4.1.0
h5py==3.8.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.25.1
huggingface-hub==0.18.0
humanize==4.8.0
Hydra==2.5
hydra-core==1.2.0
Hypercorn==0.15.0
hyperframe==6.0.1
idna==3.3
imageio==2.33.0
importlib-metadata==6.0.0
importlib-resources==5.9.0
inflate64==1.0.0
iniconfig==1.1.1
inquirer==3.1.3
ipykernel==6.19.4
ipython==7.30.1
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.8
jaxlib==0.4.7
jedi==0.18.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.1.0
json5==0.9.11
jsonlines==4.0.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.2
jupyter_server==2.0.6
jupyter_server_terminals==0.4.3
jupyterlab==3.5.2
jupyterlab-pygments==0.2.2
jupyterlab_server==2.18.0
kaggle==1.5.12
kaleido==0.2.1
keras==2.12.0
kiwisolver==1.4.4
lazy_loader==0.3
libclang==16.0.0
lightning==2.0.6
lightning-cloud==0.5.37
lightning-utilities==0.9.0
Markdown==3.3.6
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.3
mdurl==0.1.2
mistune==2.0.4
ml-collections==0.1.1
ml-dtypes==0.1.0
mongoengine==0.24.2
motor==3.3.2
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multivolumefile==0.2.3
mypy-extensions==1.0.0
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
networkx==3.1
neural-tangents==0.6.2
notebook==6.5.2
notebook_shim==0.2.2
numexpr==2.8.6
numpy==1.23.5
oauthlib==3.1.1
omegaconf==2.2.3
opencv-python-headless==4.8.1.78
opt-einsum==3.3.0
ordered-set==4.1.0
packaging==23.0
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.0
pathtools==0.1.2
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
platformdirs==2.6.2
plotly==5.14.1
pluggy==1.0.0
pprintpp==0.4.0
priority==2.0.0
prometheus-client==0.15.0
promise==2.3
prompt-toolkit==3.0.23
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
py==1.11.0
py-cpuinfo==9.0.0
py7zr==0.20.8
pyarrow==12.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybcj==1.0.2
pycparser==2.21
pycryptodomex==3.19.0
pydantic==2.0.3
pydantic_core==2.3.0
pydeck==0.8.1b0
Pygments==2.15.1
PyJWT==2.8.0
pymongo==4.6.0
Pympler==1.0.1
pyparsing==3.0.6
pyppmd==1.1.0
pyrsistent==0.19.3
pytest==6.2.5
python-dateutil==2.8.2
python-editor==1.0.4
python-json-logger==2.0.4
python-multipart==0.0.6
python-slugify==5.0.2
pytorch-lightning==2.0.6
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==5.4.1
pyzmq==24.0.1
pyzstd==0.15.9
rarfile==4.1
readchar==4.0.5
regex==2023.10.3
requests==2.28.1
requests-oauthlib==1.3.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.5
rsa==4.8
s3transfer==0.7.0
safetensors==0.4.0
scikit-image==0.21.0
scikit-learn==1.0.1
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.0
sentry-sdk==1.11.1
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sse-starlette==0.10.3
sseclient-py==1.8.0
starlette==0.27.0
starsessions==1.3.0
statsmodels==0.14.0
strawberry-graphql==0.138.1
streamlit==1.22.0
sympy==1.11.1
tables==3.8.0
tabulate==0.9.0
taskgroup==0.0.0a4
tenacity==8.2.2
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.12.0
tensorflow-datasets==4.9.2
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
tensorflow-metadata==1.13.1
termcolor==2.2.0
terminado==0.17.1
text-unidecode==1.3
texttable==1.7.0
tf2jax==0.3.4
threadpoolctl==3.0.0
tifffile==2023.7.10
timm==0.9.8
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchmetrics==1.0.1
torchvision==0.15.2
tornado==6.2
tqdm==4.62.3
traitlets==5.8.0
typeguard==4.1.5
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3
universal-analytics-python3==1.1.1
uri-template==1.2.0
urllib3==1.26.13
uvicorn==0.23.1
validators==0.20.0
voxel51-eta==0.12.0
wandb==0.15.5
watchdog==3.0.0
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
websockets==11.0.3
Werkzeug==2.0.2
wget==3.2
wrapt==1.14.1
wsproto==1.2.0
xmltodict==0.13.0
yapf==0.32.0
yarl==1.8.2
zipp==3.6.0
